package(refactor): #55 reorganize test suite structure
- Move test fixtures to tests/fixtures/ (context, main, tools, prompts, models)
- Group http_mcp tests under tests/http_mcp/ mirroring source layout
- Merge related tests: tools (4 files → 1), prompts (3 files → 1),
  server (3 files → 1), dissolve test_coverage_gaps into proper modules
- Move integration tests to tests/integration/
- Fix test_studio_transport.py typo (studio → stdio)
- Convert README security tables to bullet lists (mdformat-compatible)
- 194 tests, 94.10% coverage, all checks pass
