+++ tests/fixtures/prompts.py
@@ -4,7 +4,7 @@ from starlette.authentication import has_required_scope
 from http_mcp._mcp_types.content import TextContent
 from http_mcp._mcp_types.prompts import PromptMessage
 from http_mcp.types import Arguments, NoArguments, Prompt
-from tests.app.tools import TOOLS
+from tests.fixtures.tools import TOOLS
 
 
 class GetAdvice(BaseModel):
