+++ tests/http_mcp/_json_rcp_types/test_errors.py
