+++ tests/test_smoke.py
